feat(dashboard): Dashboard subscription, plugins beta
