+++ apps/dashboard/package.json
@@ -58,7 +58,7 @@
     "@graphql-codegen/typescript-react-apollo": "^3.2.14",
     "@iconify/react": "^3.2.1",
     "@next/bundle-analyzer": "^12.1.6",
-    "@swc/core": "^1.2.222",
+    "@swc/core": "^1.2.242",
     "@tailwindcss/forms": "^0.5.2",
     "@tailwindcss/line-clamp": "^0.4.0",
     "@tailwindcss/typography": "^0.5.4",
