+++ apps/dashboard/src/components/dashboard/pages/overview.tsx
@@ -1,14 +1,17 @@
+import dynamic from 'next/dynamic'
 import { useEffect, useState } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 import toast from 'react-hot-toast'
 
-import {
-  Guild,
-  useUpdateGuildByIdMutation,
-} from '../../../graphql/generated/schema'
+import { useUpdateGuildByIdMutation } from '../../../graphql/generated/schema'
 import { useCurrentGuildConfig } from '../../../hooks'
 import { PluginToggle } from '../../forms'
-import GlobalSavePopup from '../global-save-popup'
+
+// global save popup dynamic import
+
+const GlobalSavePopup = dynamic(() => import('../global-save-popup'), {
+  ssr: false,
+})
 
 const OverviewPluginsPage = () => {
   const { guild } = useCurrentGuildConfig()
@@ -74,24 +77,28 @@ const OverviewPluginsPage = () => {
     toast.dismiss('dashboard-global-save')
 
     if (isDirty)
-      toast.custom(
-        t => (
-          <GlobalSavePopup
-            t={t}
-            loading={isSubmitting || loading}
-            error={error}
-            successful={isSubmitSuccessful}
-            handleSubmit={saveConfig}
-            reset={reset}
-            isDirty={isDirty}
-          />
-        ),
-        {
-          duration: Infinity,
-          position: 'bottom-center',
-          id: 'dashboard-global-save',
-        },
-      )
+      setTimeout(() => {
+        toast.custom(
+          t => (
+            <GlobalSavePopup
+              t={t}
+              loading={isSubmitting || loading}
+              error={error}
+              successful={isSubmitSuccessful}
+              handleSubmit={saveConfig}
+              reset={reset}
+              isDirty={isDirty}
+            />
+          ),
+          {
+            duration: Infinity,
+            position: 'bottom-center',
+            id: 'dashboard-global-save',
+          },
+        )
+      }, 1000)
+
+    return
   }, [isDirty])
 
   return (
