+++ apps/dashboard/src/hooks/index.ts
@@ -7,3 +7,5 @@ export { useModal } from './useModal'
 export { useOnClickOutside } from './useOnClickOutside'
 export { useCurrentGuildConfig } from './useCurrentGuildConfig'
 export { useHandlePageLoading } from './useHandlePageLoading'
+export { useCurrentGuildId } from './useCurrentGuildId'
+export { useSubscribeGuildConfig } from './useSubscribeGuildConfig'
