+++ apps/dashboard/src/libs/fetchUserIcon.ts
@@ -1,0 +1,8 @@
+export const FetchUserIcon = (
+  discordId: string,
+  avatar: string | null | undefined,
+) => {
+  if (!avatar) return '/basic_discord_logo.png'
+
+  return `${process.env.NEXT_PUBLIC_DISCORD_CDN}/avatars/${discordId}/${avatar}.png?size=128`
+}
