+++ apps/dashboard/src/pages/_document.tsx
@@ -10,14 +10,8 @@ class MyDocument extends Document {
 
           <meta charSet="utf-8" />
 
-          <link rel="preconnect" href="https://fonts.googleapis.com" />
           <link
-            rel="preconnect"
-            href="https://fonts.gstatic.com"
-            crossOrigin="anonymous"
-          />
-          <link
-            href="https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;300;400;500;600;700&display=swap"
+            href="https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;300;400;500;600;700&display=optional"
             rel="stylesheet"
           />
         </Head>
