+++ apps/dashboard/src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const APOLLO_STATE_PROP_NAME = '__APOLLO_STATE__'
